- Merge `origin/main` (up to #21546) into the `performance` branch.
- Conflicts resolved by taking main's finalized form where the perf
branch was behind (`ExistenceFilterVersion`→1 per #21164,
`mdbx-go`→v0.40.1, `merge.go` `findMergeRangeInFiles` refactor, dropped
`erigon-snapshot` module dep, fusefilter deferred-close refactor,
`Versions.MustSupport`, atomic per-key prune throttle).
- Adopted main's collation-at-tip design (`CollateAndPrune` in the FCU
path, #21398/#21415) and removed the perf branch's older
`frozenBlocks`-gating (`SetFrozenBlocksProvider`/`MaxCollatableTxNum`,
`db/services/snapshot_progress.go`, its gating tests, and callers).
- Verified: `make erigon integration` build, `make lint` (clean), `make
test-short` (green).

+++ execution/cache/generic_cache.go
@@ -221,6 +221,6 @@ func (c *GenericCache[T]) PrintStatsAndReset(name string) {
 	log.Debug(name+" cache stats",
 		"hits", hits, "misses", misses, "hit_rate", hitRate,
 		"entries", c.data.Len(), "size_mb", sizeBytes/(1024*1024),
-		"capacity_mb", c.capacityB/datasize.MB, "usage_pct", usagePct,
+		"capacity_mb", int64(c.capacityB/datasize.MB), "usage_pct", usagePct,
 	)
 }
